+++ protobuf/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Improve performance of `GeneratedMessage.deepCopy`. ([#742])
 * Fix unknown enum handling in `GeneratedMessage.mergeFromProto3Json` when
-  `ignoreUnknownFields` optional arguments is `true`. ([#853])
+  the `ignoreUnknownFields` optional argument is `true`. ([#853])
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#853]: https://github.com/google/protobuf.dart/pull/853