+++ protobuf/CHANGELOG.md
@@ -3,9 +3,11 @@
 * Improve performance of `GeneratedMessage.deepCopy`. ([#742])
 * Fix unknown enum handling in `GeneratedMessage.mergeFromProto3Json` when
   the `ignoreUnknownFields` optional argument is `true`. ([#853])
+* Add `BuilderInfo` methods to support protoc-plugin 23.0.0. ([#1047])
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#853]: https://github.com/google/protobuf.dart/pull/853
+[#1047]: https://github.com/google/protobuf.dart/pull/1047
 
 ## 4.2.0