+++ protobuf/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix unknown enum handling in `GeneratedMessage.mergeFromProto3Json` when
   the `ignoreUnknownFields` optional argument is `true`. ([#853])
 * Add `BuilderInfo` methods to support protoc-plugin 23.0.0. ([#1047])
+* Generalize argument type of `PbList.from` from `List<T>` to `Iterable<T>`.
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#853]: https://github.com/google/protobuf.dart/pull/853