+++ protobuf/CHANGELOG.md
@@ -5,10 +5,12 @@
   the `ignoreUnknownFields` optional argument is `true`. ([#853])
 * Add `BuilderInfo` methods to support protoc-plugin 23.0.0. ([#1047])
 * Generalize argument type of `PbList.from` from `List<T>` to `Iterable<T>`.
+  ([#1054])
 
 [#742]: https://github.com/google/protobuf.dart/pull/742
 [#853]: https://github.com/google/protobuf.dart/pull/853
 [#1047]: https://github.com/google/protobuf.dart/pull/1047
+[#1054]: https://github.com/google/protobuf.dart/pull/1054
 
 ## 4.2.0